backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
sqlalchemy>=2.0
pydantic-settings>=2.0
email-validator>=2.0
python-jose[cryptography]>=3.3
cryptography>=41.0
python-multipart>=0.0.7
httpx>=0.25

# test
pytest>=7.0
